Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
scikit-learn==1.3.2
catboost==1.2.2
python-multipart==0.0.6
joblib==1.3.2
python-jose[cryptography]==3.3.0  # JWT token işlemleri için
passlib[bcrypt]==1.7.4           # Şifreleme için
python-dotenv==1.0.0             # Çevresel değişkenler için
aiofiles==23.2.1                # Asenkron dosya işlemleri için
tensorflow==2.15.0              # CNN modeli için
Pillow==10.2.0                  # Görüntü işleme için